+++ forward_engineering/config/constants.js
@@ -1,0 +1,5 @@
+const DEFAULT_INDENT = '    ';
+
+module.exports = {
+	DEFAULT_INDENT,
+};
