--- reverse_engineering/appDependencies.js
@@ -1,10 +1,0 @@
-let dependencies = {};
-
-const setDependencies = app => {
-	dependencies.lodash = app.require('lodash');
-};
-
-module.exports = {
-	setDependencies,
-	dependencies,
-};
